Fixed bug on multi-hop retransmit time.

+++ flopsync_v3/flooder_root_node.h
@@ -72,7 +72,7 @@ public:
     /**
      * \return the timestamp received through the radio
      */
-    unsigned int getRadioTimestamp() const { return frameStart; }
+    unsigned long long getRadioTimestamp() const { return frameStart; }
     #endif //SEND_TIMESTAMPS
 
     ~FlooderRootNode();
